fs/dcache: Fix indefinite wait in d_invalidate()
In the path of task exit, the proc dentries corresponding to this
task will be killed by moving it to a shrink list. If those dentries
are already claimed by another task for shrinking, the exiting
task waits in a tight loop until those dentries are killed. This can
potentially result in a deadlock if those dentries are corresponding
to an RT task but the task which is shrinking those dentries is a
lower priority task. Fix this by not doing tight loop, if our dentries
are claimed by other task.

Change-Id: If6848521469db7dea2bbba0dbaf8597094716267
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>