ARM: dts: msm: Specify proper address and size cells for emac on mdm9607
The phy subnode of this emac node has a single cell, indicating
an address of some kind. Set address-cells to 1 and size-cells to
0 to indicate this to dtc so that dtc doesn't print a warning
about incorrect reg properties for the phy node.

The address-cells was specificd as 0 so that the interrupt-map
could be used, but we have the 'interrupts-extended' properties
to support this now so that we don't have to make this node into
an interrupt controller just to remap interrupts to real
interrupt controllers.

Change-Id: I3669f361feecac9747447512bf571939afbbf95d
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>