usb: gsi: Don't disable endpoints as part of flow control
Currently GSI endpoints are disabled when host enables
flow control for RNDIS function. When these endpoints are
enabled again as part of flow control disable, no transactions
are observed on these endpoints. Fix this by not disabling the
endpoints as part of flow control enable but just stop the
active transfers on these endpoints.

Change-Id: I391a7048188b2a63e2df993fcebf7a6e78eaef14
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>